Support for Pods as a sync hook Add app label to pod metadata in job.spec.template.labels Fix issue where hooks could bypass a project whitelist/blacklist Fix issue where deletion of hooks did not perform a cascaded deletion
